\name{mhca-package}
\alias{mhca-package}
\alias{mhca}
\docType{package}
\title{Mahalanobis HCA}
\description{Mahalanobis distance-based hierarchical cluster analysis,
in which elliptical clusters get found naturally.}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{mhca}
\packageIndices{mhca}
Maintainer: Tomas Sieger <tomas.sieger@seznam.cz>
}
\author{Tomas Sieger, Karel Fiser}
\keyword{ package }
